+++ packages/lumx-vue/src/components/text-field/index.ts
@@ -1,0 +1,1 @@
+export { default as RawInputText, type RawInputTextProps } from './RawInputText';