+++ packages/lumx-vue/src/index.ts
@@ -22,6 +22,7 @@ export * from './components/progress';
 export * from './components/radio-button';
 export * from './components/skeleton';
 export * from './components/switch';
+export * from './components/text-field';
 export * from './components/table';
 export * from './components/text';
 export * from './components/thumbnail';